Moving cast utils to its own module
