+++ nemo/collections/asr/models/asr_model.py
@@ -20,8 +20,7 @@ import torch
 from nemo.core.classes import ModelPT
 from nemo.core.classes.exportable import Exportable
 from nemo.core.classes.mixins import AccessMixin
-from nemo.utils import logging, model_utils
-from nemo.utils.export_utils import cast_all
+from nemo.utils import cast_all, logging, model_utils
 
 __all__ = ['ASRModel']
